chore: introduce data-list-component helper
